Move platform specific assets to Game properties
